refactor: fix collection list (#9423)
* refactor: naming stuff

* refactor: naming stuff
