+++ packages/plugins/Solana/src/state/Hub/hub.ts
@@ -6,7 +6,7 @@ import {
     SolanaFungible,
     SolanaNonFungible,
     NFTScanNonFungibleTokenSolana,
-    SolanaSimpleHash,
+    SimpleHashSolana,
 } from '@masknet/web3-providers'
 import type { FungibleTokenAPI, NonFungibleTokenAPI, PriceAPI } from '@masknet/web3-providers/types'
 import {
@@ -64,9 +64,9 @@ class HubNonFungibleClient extends HubStateNonFungibleClient<ChainId, SchemaType
                 [SourceType.MagicEden]: MagicEden,
                 [SourceType.Solana]: SolanaNonFungible,
                 [SourceType.NFTScan]: NFTScanNonFungibleTokenSolana,
-                [SourceType.SimpleHash]: SolanaSimpleHash,
+                [SourceType.SimpleHash]: SimpleHashSolana,
             },
-            [NFTScanNonFungibleTokenSolana, MagicEden, SolanaNonFungible, SolanaSimpleHash],
+            [NFTScanNonFungibleTokenSolana, MagicEden, SolanaNonFungible, SimpleHashSolana],
             initial,
         )
     }
