+++ packages/web3-providers/src/entry.ts
@@ -64,11 +64,10 @@ import { SpaceID_API } from './SpaceID/index.js'
 import { BonfidaAPI } from './Bonfida/index.js'
 import { RedPacketAPI } from './RedPacket/index.js'
 import { SmartPayOwnerAPI } from './SmartPay/apis/OwnerAPI.js'
-import { SimpleHashProviderAPI } from './SimpleHash/index.js'
+import { SimpleHashAPI_EVM, SimpleHashAPI_Solana } from './SimpleHash/index.js'
 import { RSS3API } from './RSS3/index.js'
 import { LidoAPI } from './Lido/index.js'
 import { Web3BioAPI } from './Web3Bio/index.js'
-import { SolanaSimpleHashProvider } from './SimpleHash/SolanaSimpleHashProvider.js'
 
 export const OpenSea = new OpenSeaAPI()
 export const Lido = new LidoAPI()
@@ -100,8 +99,8 @@ export const Sentry = new SentryAPI()
 export const TheGraphDomain = new TheGraphDomainAPI()
 export const RedPacket = new RedPacketAPI()
 export const TheGraphRedPacket = new TheGraphRedPacketAPI()
-export const SimpleHash = new SimpleHashProviderAPI()
-export const SolanaSimpleHash = new SolanaSimpleHashProvider()
+export const SimpleHashEVM = new SimpleHashAPI_EVM()
+export const SimpleHashSolana = new SimpleHashAPI_Solana()
 
 // Wallet
 export {
